[refactoring] moving annotation
