+++ src/main/java/edu/vt/codewaveservice/processor/CriticalProcessor.java
@@ -1,4 +1,4 @@
-package edu.vt.codewaveservice.processor.impl;
+package edu.vt.codewaveservice.processor;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
